+++ modules/Grades/ProgressReports.php
@@ -70,7 +70,7 @@ if ( $_REQUEST['modfunc'] === 'save' )
 	//$LO_columns += array('POINTS' => _('Points'),'PERCENT_GRADE' => _('Percent'),'LETTER_GRADE' => _('Letter'),'COMMENT' => _('Comment'));
 	$LO_columns += [ 'COMMENT' => _( 'Comment' ) ];
 
-	$extra2['SELECT_ONLY'] = "ga.TITLE,ga.ASSIGNED_DATE,ga.DUE_DATE,gt.ASSIGNMENT_TYPE_ID,gg.POINTS,ga.POINTS AS TOTAL_POINTS,gt.FINAL_GRADE_PERCENT,gg.COMMENT,gg.POINTS AS PERCENT_GRADE,gg.POINTS AS LETTER_GRADE,CASE WHEN (ga.ASSIGNED_DATE IS NULL OR CURRENT_DATE>=ga.ASSIGNED_DATE) AND (ga.DUE_DATE IS NULL OR CURRENT_DATE>=ga.DUE_DATE OR CURRENT_DATE>(SELECT END_DATE FROM school_marking_periods WHERE MARKING_PERIOD_ID=ga.MARKING_PERIOD_ID)) THEN 'Y' ELSE NULL END AS DUE,gt.TITLE AS CATEGORY_TITLE";
+	$extra2['SELECT_ONLY'] = "ga.ASSIGNMENT_ID,ga.TITLE,ga.ASSIGNED_DATE,ga.DUE_DATE,gt.ASSIGNMENT_TYPE_ID,gg.POINTS,ga.POINTS AS TOTAL_POINTS,gt.FINAL_GRADE_PERCENT,gg.COMMENT,gg.POINTS AS PERCENT_GRADE,gg.POINTS AS LETTER_GRADE,ga.WEIGHT,CASE WHEN (ga.ASSIGNED_DATE IS NULL OR CURRENT_DATE>=ga.ASSIGNED_DATE) AND (ga.DUE_DATE IS NULL OR CURRENT_DATE>=ga.DUE_DATE OR CURRENT_DATE>(SELECT END_DATE FROM school_marking_periods WHERE MARKING_PERIOD_ID=ga.MARKING_PERIOD_ID)) THEN 'Y' ELSE NULL END AS DUE,gt.TITLE AS CATEGORY_TITLE";
 
 	$extra2['FROM'] = '';
 
@@ -178,7 +178,7 @@ if ( $_REQUEST['modfunc'] === 'save' )
 
 			$extra['WHERE'] .= " AND ss.COURSE_PERIOD_ID='" . (int) $cp_id . "'";
 
-			$student_points = $total_points = $percent_weights = [];
+			$student_points = $weighted_grade = $total_points = $percent_weights = $total_weights = [];
 
 			$grades_RET = GetStuList( $extra );
 
@@ -191,7 +191,7 @@ if ( $_REQUEST['modfunc'] === 'save' )
 
 			$gradebook_config = ProgramUserConfig( 'Gradebook', $teacher_id );
 
-			$sum_student_points = $sum_total_points = $sum_points = $sum_percent = 0;
+			$sum_student_points = $sum_weighted_grade = $sum_total_points = $sum_points = $sum_percent = $sum_weights = 0;
 
 			foreach ( (array) $percent_weights as $assignment_type_id => $percent )
 			{
@@ -199,15 +199,53 @@ if ( $_REQUEST['modfunc'] === 'save' )
 				$sum_total_points += $total_points[$assignment_type_id];
 				$sum_points += $student_points[$assignment_type_id] * ( ! empty( $gradebook_config['WEIGHT'] ) && $percent ? $percent / $total_points[$assignment_type_id] : 1 );
 				$sum_percent += ( ! empty( $gradebook_config['WEIGHT'] ) && $percent ? $percent : $total_points[$assignment_type_id] );
+
+				// @since 11.0 Add Weight Assignments option
+				$sum_weighted_grade += ( ! empty( $gradebook_config['WEIGHT'] ) && $percent ?
+					$percent * $weighted_grade[$assignment_type_id] :
+					$weighted_grade[$assignment_type_id] );
+
+				$sum_weights += ( ! empty( $gradebook_config['WEIGHT'] ) && $percent ?
+					$percent * $total_weights[$assignment_type_id] :
+					$total_weights[$assignment_type_id] );
 			}
+
+			$sum_grade = 0;
 
 			if ( $sum_percent > 0 )
 			{
-				$sum_points /= $sum_percent;
+				$sum_grade = $sum_points / $sum_percent;
 			}
-			else
+
+			if ( ! empty( $gradebook_config['WEIGHT_ASSIGNMENTS'] ) )
 			{
-				$sum_points = 0;
+				// @since 11.0 Add Weight Assignments option
+				if ( $sum_weights > 0 )
+				{
+					$sum_grade = $sum_weighted_grade / $sum_weights;
+				}
+
+				foreach ( $grades_RET as $assignment_type_id => $grades )
+				{
+					if ( isset( $_REQUEST['by_category'] )
+						&& $_REQUEST['by_category'] == 'Y' )
+					{
+						foreach ( $grades as $grade_i => $grade )
+						{
+							$assignment_weight = ' <span class="size-1">(' .
+								_( 'Weight' ) . ' ' . (int) $grades['WEIGHT'] . ')</span>';
+
+							$grades_RET[$assignment_type_id][$grade_i]['TITLE'] .= $assignment_weight;
+						}
+					}
+					else
+					{
+						$assignment_weight = ' <span class="size-1">(' .
+							_( 'Weight' ) . ' ' . (int) $grades['WEIGHT'] . ')</span>';
+
+						$grades_RET[$assignment_type_id]['TITLE'] .= $assignment_weight;
+					}
+				}
 			}
 
 			if ( isset( $_REQUEST['by_category'] )
@@ -228,6 +266,14 @@ if ( $_REQUEST['modfunc'] === 'save' )
 					$type_percent = ! empty( $total_points[$assignment_type_id] ) ?
 						$student_points[$assignment_type_id] / $total_points[$assignment_type_id] :
 						'';
+
+					if ( ! empty( $gradebook_config['WEIGHT_ASSIGNMENTS'] ) )
+					{
+						// @since 11.0 Add Weight Assignments option
+						$type_percent = ! empty( $total_weights[$assignment_type_id] ) ?
+							$weighted_grade[$assignment_type_id] / $total_weights[$assignment_type_id] :
+							'';
+					}
 
 					$percent_grade = $letter_grade = '&nbsp;';
 
@@ -254,7 +300,7 @@ if ( $_REQUEST['modfunc'] === 'save' )
 				}
 			}
 
-			$percent = _makeLetterGrade( $sum_points, $cp_id, $teacher_id, '%' );
+			$percent = _makeLetterGrade( $sum_grade, $cp_id, $teacher_id, '%' );
 
 			// Do not add Total to $link['add']['html']: PDF and no AllowEdit().
 			$total_last_row = [
@@ -263,7 +309,7 @@ if ( $_REQUEST['modfunc'] === 'save' )
 				'DUE_DATE' => '&nbsp;',
 				'POINTS' => '<b>' . $sum_student_points . '&nbsp;/&nbsp;' . $sum_total_points . '</b>',
 				'PERCENT_GRADE' => '<b>' . _Percent( $percent ) . '</b>',
-				'LETTER_GRADE' => '<b>' . _makeLetterGrade( $sum_points, $cp_id, $teacher_id ) . '</b>',
+				'LETTER_GRADE' => '<b>' . _makeLetterGrade( $sum_grade, $cp_id, $teacher_id ) . '</b>',
 				'COMMENT' => '&nbsp;',
 			];
 
@@ -396,7 +442,7 @@ if ( ! $_REQUEST['modfunc'] )
  */
 function _makeExtraPoints( $value, $column )
 {
-	global $THIS_RET, $student_points, $total_points, $percent_weights;
+	global $THIS_RET, $student_points, $total_points, $percent_weights, $total_weights;
 
 	if ( $THIS_RET['TOTAL_POINTS'] == '0' )
 	{
@@ -433,6 +479,14 @@ function _makeExtraPoints( $value, $column )
 		$total_points[$THIS_RET['ASSIGNMENT_TYPE_ID']] += $THIS_RET['TOTAL_POINTS'];
 
 		$percent_weights[$THIS_RET['ASSIGNMENT_TYPE_ID']] = $THIS_RET['FINAL_GRADE_PERCENT'];
+
+		if ( ! isset( $total_weights[$THIS_RET['ASSIGNMENT_TYPE_ID']] ) )
+		{
+			$total_weights[$THIS_RET['ASSIGNMENT_TYPE_ID']] = 0;
+		}
+
+		// @since 11.0 Add Weight Assignments option
+		$total_weights[$THIS_RET['ASSIGNMENT_TYPE_ID']] += $THIS_RET['WEIGHT'];
 	}
 
 	return (float) $value . '&nbsp;/&nbsp;' . $THIS_RET['TOTAL_POINTS'];
@@ -453,7 +507,7 @@ function _makeExtraPoints( $value, $column )
  */
 function _makeExtraGrade( $value, $column )
 {
-	global $THIS_RET, $cp_id, $teacher_id;
+	global $THIS_RET, $cp_id, $teacher_id, $weighted_grade;
 
 	if ( isset( $THIS_RET['TOTAL_POINTS'] )
 		&& $THIS_RET['TOTAL_POINTS'] == '0' )
@@ -479,6 +533,14 @@ function _makeExtraGrade( $value, $column )
 	}
 
 	$percent = _makeLetterGrade( $value / $THIS_RET['TOTAL_POINTS'], $cp_id, $teacher_id, '%' );
+
+	if ( ! isset( $weighted_grade[$THIS_RET['ASSIGNMENT_TYPE_ID']] ) )
+	{
+		$weighted_grade[$THIS_RET['ASSIGNMENT_TYPE_ID']] = 0;
+	}
+
+	// @since 11.0 Add Weight Assignments option
+	$weighted_grade[$THIS_RET['ASSIGNMENT_TYPE_ID']] += ( $value / $THIS_RET['TOTAL_POINTS'] ) * $THIS_RET['WEIGHT'];
 
 	return _Percent( $percent, 2 );
 }